+++ t-rex-webserver/src/server.rs
@@ -229,7 +229,17 @@ pub fn webserver(args: ArgMatches<'static>) {
             .wrap(Cors::new().send_wildcard().allowed_methods(vec!["GET"]))
             .service(web::resource("/index.json").route(web::get().to_async(mvt_metadata)))
             .service(web::resource("/fontstacks.json").route(web::get().to(fontstacks)))
-            .service(web::resource("/fonts/{fonts}/{range}.pbf").route(web::get().to(fonts_pbf)))
+            .service(web::resource("/fonts/{fonts}/{range}.pbf").route(web::get().to(fonts_pbf)));
+        for static_dir in &static_dirs {
+            let dir = &static_dir.dir;
+            if std::path::Path::new(dir).is_dir() {
+                info!("Serving static files from directory '{}'", dir);
+                app = app.service(fs::Files::new(&static_dir.path, dir));
+            } else {
+                warn!("Static file directory '{}' not found", dir);
+            }
+        }
+        app = app
             .service(
                 web::resource("/{tileset}.style.json")
                     .route(web::get().to_async(tileset_style_json)),
@@ -242,15 +252,6 @@ pub fn webserver(args: ArgMatches<'static>) {
             .service(
                 web::resource("/{tileset}/{z}/{x}/{y}.pbf").route(web::get().to_async(tile_pbf)),
             );
-        for static_dir in &static_dirs {
-            let dir = &static_dir.dir;
-            if std::path::Path::new(dir).is_dir() {
-                info!("Serving static files from directory '{}'", dir);
-                app = app.service(fs::Files::new(&static_dir.path, dir));
-            } else {
-                warn!("Static file directory '{}' not found", dir);
-            }
-        }
         if mvt_viewer {
             app = app
                 .service(web::resource("/drilldown").route(web::get().to_async(drilldown_handler)));