+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 #### Features
 
 * [Template](https://tera.netlify.com/docs/templates/) support in configuration files
+* Load environment variables from a file named `.env` in the current directory
+  or any of its parents
 * Initialize service only once before starting HttpServer threads
 * Share DB connection pool between all HttpServer threads
 * Migrate web server to Actix web 1.0